Move common model code into library project
