+++ app-mvp/build.gradle
@@ -31,6 +31,7 @@ dependencies {
     compile dependencies.picasso
     compile dependencies.rxAndroid
     compile dependencies.circleImageView
+    compile project(':model')
 
     testCompile dependencies.jUnit
     testCompile dependencies.mockito
