+++ app-mvp/src/main/java/uk/ivanc/archimvp/view/MainActivity.java
@@ -19,9 +19,9 @@ import android.widget.TextView;
 
 import java.util.List;
 
+import uk.ivanc.archi.model.Repository;
 import uk.ivanc.archimvp.R;
 import uk.ivanc.archimvp.RepositoryAdapter;
-import uk.ivanc.archimvp.model.Repository;
 import uk.ivanc.archimvp.presenter.MainPresenter;
 
 public class MainActivity extends AppCompatActivity implements MainMvpView {
