+++ app-mvp/src/test/java/uk/ivanc/archimvp/MainPresenterTest.java
@@ -16,8 +16,8 @@ import retrofit2.Response;
 import retrofit2.adapter.rxjava.HttpException;
 import rx.Observable;
 import rx.schedulers.Schedulers;
-import uk.ivanc.archimvp.model.GithubService;
-import uk.ivanc.archimvp.model.Repository;
+import uk.ivanc.archi.model.GithubService;
+import uk.ivanc.archi.model.Repository;
 import uk.ivanc.archimvp.presenter.MainPresenter;
 import uk.ivanc.archimvp.util.MockModelFabric;
 import uk.ivanc.archimvp.view.MainMvpView;
