+++ model/build.gradle
@@ -1,15 +1,10 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 
 android {
     compileSdkVersion rootProject.ext.androidCompileSdkVersion
     buildToolsVersion rootProject.ext.androidBuildToolsVersion
 
     defaultConfig {
-        applicationId "uk.ivanc.archi"
-        minSdkVersion rootProject.ext.androidMinSdkVersion
-        targetSdkVersion rootProject.ext.androidTargetSdkVersion
-        versionCode 1
-        versionName "1.0"
     }
     buildTypes {
         release {
@@ -17,19 +12,13 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    publishNonDefault true
 }
 
 dependencies {
     Map<String, String> dependencies = rootProject.ext.dependencies;
 
-    compile dependencies.appCompat
-    compile dependencies.cardView
-    compile dependencies.recyclerView
     compile dependencies.retrofit
     compile dependencies.retrofitConverterGson
     compile dependencies.retrofitAdapterRxJava
-    compile dependencies.picasso
-    compile dependencies.rxAndroid
-    compile dependencies.circleImageView
-    compile project(':model')
 }
