+++ model/src/main/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="uk.ivanc.archi.model">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+
+    <application />
+
+</manifest>
