+++ model/src/main/java/uk/ivanc/archi/model/GithubService.java
