+++ model/src/main/java/uk/ivanc/archi/model/User.java
