+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':app-mvp', ':app-mvvm'
+include ':model', ':app', ':app-mvp', ':app-mvvm'
